Fix for timestamped indexed
This fixes an issue when using timestamped indexes via the "index_timestamp" parameter.

The index name is being set before the JODA timestamp pattern has been evaluated which means the index is created is a literal string.
e.g. index-YYYY-MM-dd instead of index-2014-06-28

To fix this, setTimeWindowed paremater is set before the index name is set in setIndex.