Fixing broken feeder mode, update to ES 1.3.1
The feeder mode was broken due to wrong classpath control in the example shell scripts
and the river state move into the cluster state. Cluster state is not possible
to use in TransportClient, so a workaround was introduced to let the JDBC plugin
start also in a TransportClient.

More comments added at example shell scripts to clarify the use of feeder mode
in a JVM outside of an Elasticsearch node.

Update to Elasticsearch 1.3.1

An NPE in the river state getting should be fixed.

Dropping unused dependency of hamcrest matchers.